mEm OnBoarding Site, Version 1
This was the first "rough" draft that I came up with, that is fully
responsive.  I was planning on adding a few nice design elements, such
as a parallax background with splashes of color/graphics and of course I
planned on sprucing up the CSS so that everything was positioned
correctly as well, however, this version was shot down because the
"sliding" function and the responsive design were not what was wanted by
others.